Refactor architecture and testing structure
- Removed ProductionDependencyBoundaryTests.cs as its functionality is now covered by ApplicationArchitectureTests.cs.
- Updated AGENTS.md to clarify ownership of database infrastructure and persistence responsibilities.
- Revised README.md to reflect changes in the cave revision and proposal architecture.
- Deleted outdated documentation on cave mutation inventory and cave revisions and review architecture.
- Introduced new documentation for cave revisions and proposals, emphasizing the separation of published and proposed states.
- Added import compatibility documentation to outline observable import semantics and approved semantic differences.
- Enhanced testing documentation to streamline test philosophy and clarify testing contracts.
- Implemented ApplicationArchitectureTests.cs to ensure application orchestration types do not directly depend on database infrastructure.
