--- Planarian.Tests.Integration/Caves/Revisions/SnapshotStorageMeasurementTests.cs
@@ -1,28 +1,0 @@
-using System.Text;
-using Npgsql;
-using Planarian.Model.Database.Entities.RidgeWalker;
-using Planarian.Model.Database.Revisions;
-using Planarian.Model.Shared.Helpers;
-using Xunit;
-using Xunit.Abstractions;
-
-namespace Planarian.Tests;
-
-public sealed class SnapshotStorageMeasurementTests(PostgresTestServer fixture,ITestOutputHelper output):IClassFixture<PostgresTestServer>
-{
-    [Fact]
-    public async Task RepresentativeSnapshotsMeasureJsonbAndRelationStorageWithoutSpeculativeTuning()
-    {
-        await using var d=await fixture.CreateDatabaseAsync(nameof(RepresentativeSnapshotsMeasureJsonbAndRelationStorageWithoutSpeculativeTuning));const string account="snapacct01";await SeedAccount(d,account);
-        var variants=new Dictionary<string,CavePublishedSnapshotV1>{["typical"]=Snapshot(account,900,8),["long-narrative"]=Snapshot(account,50000,8),["tag-heavy-cave"]=Snapshot(account,1200,100),["one-entrance"]=Snapshot(account,1000,12,1),["three-entrances"]=Snapshot(account,1000,12,3),["ten-entrances"]=Snapshot(account,1000,12,10),["tag-heavy-entrances"]=Snapshot(account,1000,12,10,20),["file-reference-heavy"]=Snapshot(account,1000,12,3,5,100)};
-        var ids=new Dictionary<string,string>();await using(var db=d.CreateDbContext("measure",account)){foreach(var p in variants){var id=IdGenerator.Generate();ids[p.Key]=id;db.CaveRevisions.Add(Revision(id,account,p.Value));}await db.SaveChangesAsync();}
-        var representative=new Dictionary<string,object>();await using(var cn=new NpgsqlConnection(d.ConnectionString)){await cn.OpenAsync();foreach(var p in variants){await using var c=new NpgsqlCommand("select pg_column_size(\"SnapshotJson\") from \"CaveRevisions\" where \"Id\"=@id",cn);c.Parameters.AddWithValue("id",ids[p.Key]);var pg=Convert.ToInt64(await c.ExecuteScalarAsync());var json=CaveSnapshotJson.Serialize(p.Value);representative[p.Key]=new{serializedBytes=Encoding.UTF8.GetByteCount(json),pgColumnSize=pg};}}
-        await using(var db=d.CreateDbContext("measure",account)){var keys=variants.Keys.ToArray();for(var i=0;i<1024;i++){var s=variants[keys[i%keys.Length]] with{CaveId=IdGenerator.Generate(),Narrative=(variants[keys[i%keys.Length]].Narrative??"")+new string((char)('a'+i%26),i%700)};db.CaveRevisions.Add(Revision(IdGenerator.Generate(),account,s));if((i+1)%256==0){await db.SaveChangesAsync();db.ChangeTracker.Clear();}}if(db.ChangeTracker.HasChanges())await db.SaveChangesAsync();}
-        object distribution;await using(var cn=new NpgsqlConnection(d.ConnectionString)){await cn.OpenAsync();await using var c=new NpgsqlCommand("select count(*),percentile_cont(.50) within group(order by pg_column_size(\"SnapshotJson\")),percentile_cont(.95) within group(order by pg_column_size(\"SnapshotJson\")),percentile_cont(.99) within group(order by pg_column_size(\"SnapshotJson\")),max(pg_column_size(\"SnapshotJson\")),pg_relation_size('\"CaveRevisions\"'),pg_total_relation_size('\"CaveRevisions\"') from \"CaveRevisions\" where \"AccountId\"=@a",cn);c.Parameters.AddWithValue("a",account);await using var r=await c.ExecuteReaderAsync();Assert.True(await r.ReadAsync());var count=r.GetInt64(0);var total=r.GetInt64(6);var per=total/(double)count;distribution=new{rowCount=count,p50=Convert.ToDouble(r.GetValue(1)),p95=Convert.ToDouble(r.GetValue(2)),p99=Convert.ToDouble(r.GetValue(3)),max=Convert.ToInt64(r.GetValue(4)),relationBytes=r.GetInt64(5),totalRelationBytes=total,projected100kBytes=per*100000,projected1mBytes=per*1000000};}
-        await using(var cn=new NpgsqlConnection(d.ConnectionString)){await cn.OpenAsync();await using var c=new NpgsqlCommand("select count(*) from pg_indexes where tablename='CaveRevisions' and indexdef ilike '%using gin%' and indexdef ilike '%SnapshotJson%'",cn);Assert.Equal(0L,Convert.ToInt64(await c.ExecuteScalarAsync()));}
-        output.WriteLine("SNAPSHOT_STORAGE_MEASUREMENTS "+System.Text.Json.JsonSerializer.Serialize(new{representative,distribution}));
-    }
-    private static CaveRevision Revision(string id,string account,CavePublishedSnapshotV1 s)=>new(){Id=id,AccountId=account,CaveId=s.CaveId,Source=CaveRevisionSource.SystemBaseline,Operation=CaveRevisionOperation.Create,SnapshotSchemaVersion=1,SnapshotJson=CaveSnapshotJson.Serialize(s)};
-    private static CavePublishedSnapshotV1 Snapshot(string account,int narrative,int caveTags,int entrances=0,int entranceTags=5,int files=0){var id=IdGenerator.Generate();var caveRoles=new[]{SnapshotTagRole.Geology,SnapshotTagRole.GeologicAge,SnapshotTagRole.MapStatus,SnapshotTagRole.PhysiographicProvince,SnapshotTagRole.Archeology,SnapshotTagRole.Biology,SnapshotTagRole.CaveOther,SnapshotTagRole.Cartographer,SnapshotTagRole.CaveReportedBy};var entRoles=new[]{SnapshotTagRole.EntranceStatus,SnapshotTagRole.EntranceHydrology,SnapshotTagRole.FieldIndication,SnapshotTagRole.EntranceReportedBy,SnapshotTagRole.EntranceOther};return new CavePublishedSnapshotV1{CaveId=id,AccountId=account,Name="Representative Cave",AlternateNames=["One","Two"],State=new SnapshotReference("state00001","Tennessee",null,"TN"),County=new SnapshotReference("county0001","Representative County","REP"),CountyNumber=123,LengthFeet=4200,DepthFeet=380,MaxPitDepthFeet=160,NumberOfPits=7,Narrative=new string('n',narrative),ReportedOn=new DateTime(2026,8,1,0,0,0,DateTimeKind.Utc),Tags=Enumerable.Range(0,caveTags).Select(i=>new SnapshotTagReference(caveRoles[i%caveRoles.Length],$"ct{i:D8}"[..10],$"Cave tag {i:D3}")).ToList(),Entrances=Enumerable.Range(0,entrances).Select(i=>new CaveEntranceSnapshotV1{Id=$"en{i:D8}"[..10],Name=$"Entrance {i}",IsPrimary=i==0,Description=new string('d',180+i*7),Latitude=35.1+i/10000d,Longitude=-86.2-i/10000d,Elevation=600+i*10,Srid=4326,LocationQualityTagId="quality001",LocationQualityNameAtRevision="Survey Grade",Tags=Enumerable.Range(0,entranceTags).Select(j=>new SnapshotTagReference(entRoles[j%entRoles.Length],$"et{j:D8}"[..10],$"Entrance tag {j:D3}")).ToList()}).ToList(),Files=Enumerable.Range(0,files).Select(i=>new CaveFileSnapshotV1{Id=$"fl{i:D8}"[..10],FileTypeTagId="filetype01",FileTypeNameAtRevision="Survey Document",FileName=$"survey-{i:D4}.pdf",DisplayName=$"Survey {i:D4}"}).ToList()};}
-    private static async Task SeedAccount(PostgresTestDatabase d,string account){await using var c=new NpgsqlConnection(d.ConnectionString);await c.OpenAsync();await using var q=c.CreateCommand();q.CommandText="insert into \"Accounts\"(\"Id\",\"Name\",\"CountyIdDelimiter\",\"DefaultViewAccessAllCaves\",\"ExportEnabled\",\"CreatedOn\") values(@a,'Snapshot Measurement','-',false,true,now())";q.Parameters.AddWithValue("a",account);await q.ExecuteNonQueryAsync();}
-}
